+++ angular-app/src/app/app-routing.module.ts
@@ -15,6 +15,32 @@ const routes: Routes = [
     component: PocDetailComponent,
   },
   {
+    path: 'gana-puntos',
+    loadChildren: () => import('./gana-puntos/gana-puntos.module').then(m => m.GanaPuntosModule),
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'comercios-aliados',
+    loadChildren: () => import('./gana-puntos/comercios-aliados/comercios-aliados.module').then(m => m.ComerciosAliadosModule),
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'actualizar-cedula',
+    loadChildren: () => import('./gana-puntos/actualizar-cedula/actualizar-cedula.module').then(m => m.ActualizarCedulaModule),
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'escanea-ya',
+    loadChildren: () => import('./gana-puntos/escanea-ya/escanea-ya.module').then(m => m.EscaneaYaModule),
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'como-redimir',
+    loadChildren: () => import('./gana-puntos/como-redimir/como-redimir.module').then(m => m.ComoRedimirModule),
+    canActivate: [AuthGuard]
+  },
+
+  {
     path: 'auth',
     loadChildren: () => import('./auth/auth.module').then(m => m.AuthModule),
     canActivate: [AuthGuard]