+++ angular-app/src/app/gana-puntos/comercios-aliados/comercios-aliados.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-comercios-aliados',
+  templateUrl: './comercios-aliados.component.html',
+  styleUrls: ['./comercios-aliados.component.scss']
+})
+export class ComerciosAliadosComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}