+++ angular-app/src/app/gana-puntos/como-redimir/como-redimir.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-como-redimir',
+  templateUrl: './como-redimir.component.html',
+  styleUrls: ['./como-redimir.component.scss']
+})
+export class ComoRedimirComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}