+++ angular-app/src/app/gana-puntos/escanea-ya/escanea-ya.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-escanea-ya',
+  templateUrl: './escanea-ya.component.html',
+  styleUrls: ['./escanea-ya.component.scss']
+})
+export class EscaneaYaComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}