+++ angular-app/src/app/gana-puntos/gana-puntos.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-gana-puntos',
+  templateUrl: './gana-puntos.component.html',
+  styleUrls: ['./gana-puntos.component.scss']
+})
+export class GanaPuntosComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}